+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/job/JobSpec.scala
@@ -21,6 +21,8 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.dimajix.common.TypeRegistry
 import com.dimajix.flowman.common.ParserUtils.splitSettings
 import com.dimajix.flowman.execution.Context
+import com.dimajix.flowman.execution.Phase
+import com.dimajix.flowman.execution.PhaseExecutionPolicy
 import com.dimajix.flowman.model.Category
 import com.dimajix.flowman.model.Job
 import com.dimajix.flowman.model.JobIdentifier
@@ -59,13 +61,14 @@ object JobSpec extends TypeRegistry[JobSpec] {
 }
 
 final class JobSpec extends NamedSpec[Job] {
-    @JsonProperty(value="extends") private var parents:Seq[String] = Seq()
+    @JsonProperty(value="extends") private var parents:Seq[String] = Seq.empty
     @JsonProperty(value="description") private var description:Option[String] = None
-    @JsonProperty(value="parameters") private var parameters:Seq[JobSpec.Parameter] = Seq()
-    @JsonProperty(value="environment") private var environment: Seq[String] = Seq()
-    @JsonProperty(value="targets") private var targets: Seq[String] = Seq()
+    @JsonProperty(value="parameters") private var parameters:Seq[JobSpec.Parameter] = Seq.empty
+    @JsonProperty(value="environment") private var environment: Seq[String] = Seq.empty
+    @JsonProperty(value="targets") private var targets: Seq[String] = Seq.empty
     @JsonProperty(value="metrics") private var metrics:Option[MetricBoardSpec] = None
-    @JsonProperty(value="hooks") private var hooks: Seq[HookSpec] = Seq()
+    @JsonProperty(value="hooks") private var hooks: Seq[HookSpec] = Seq.empty
+    @JsonProperty(value="phases") private var phases: Map[String,String] = Map.empty
 
     override def instantiate(context: Context, properties:Option[Job.Properties] = None): Job = {
         require(context != null)
@@ -77,7 +80,8 @@ final class JobSpec extends NamedSpec[Job] {
             splitSettings(environment).toMap,
             targets.map(context.evaluate).map(TargetIdentifier.parse),
             metrics,
-            hooks
+            hooks,
+            phases.toSeq.map(kv => Phase.ofString(kv._1) -> PhaseExecutionPolicy.ofString(kv._2)).toMap
         )
 
         Job.merge(job, parents)