+++ docs/releases.md
@@ -25,10 +25,12 @@ changes over time.
 * github-289: Color log output via log4j configuration (requires log4j 2.x)
 * Bump postgresql from 42.4.1 to 42.4.3 in /flowman-plugins/postgresql
 * Bump loader-utils from 1.4.0 to 1.4.2
-* github-293: Fatal exceptions in parallel mapping instantiation cause deadlock
+* github-293: [BUG] Fatal exceptions in parallel mapping instantiation cause deadlock
 * github-273: Support projects contained in (fat) jar files
-* github-294: Parallel execution should not execute more targets after errors
+* github-294: [BUG] Parallel execution should not execute more targets after errors
 * github-295: Create build profile for CDP 7.1 with Spark 3.2
+* github-296: Update npm dependencies (vuetify & co)
+* github-297: Parametrize when to execute a specific phase
 
 
 ### Version 0.29.0 - 2022-11-08